Delete probe file on edit, use mtime for filename
This fixes very specific situations where partially corrupted/failed
edits followed by successful attempts may result in a file that has
changed in some way, but still has the same size and timestamps,
resulting in a stale probe file being used.

Changing the probe file's filename from using `st_ctime` to `st_mtime`
also assists with this by making it less likely for files tampered with
outside of PyPlayer to not have their probe files refreshed.

Additionally, `self.open_probe_file` now has a `verbose` parameter for
controling whether a message appears on the statusbar if `file` does
not exist.

Note: This is not used for concatenating in this commit, but will be
added in the coming commits.