Optimized drag-and-drops, improved subtitle drops
`QVideoPlayer.dragdrop_files`, `QVideoPlayer.dragdrop_subtitle_count`,
and `QVideoPlayer.dragdrop_is_folder` have been added to optimize drag-
and-drop events and allow for even clearer hints on the player,
detecting and displaying whether or not you're about to drop one
subtitle file, multiple subtitle files, a combination of subtitles
and media, or if the drop you're about to do will be ignored.

Combination drops are now supported. The first valid media file in a
drop is opened, then all valid subtitle files in the drop are added
after, if any media is playing (regardless of whether or not the media
currently playing is the one that was dropped).

Additionally, a handful of regression-crashes with drag-and-drops have
been fixed, and `self.browse_for_subtitle_files` has had its subtitle-
adding code split into a new method (`self.add_subtitle_files()`) for
clarity.