Added an operation counter for the current save
This adds the properties `SaveParameters.operation_count` and
`SaveParameters.operations_completed` to allow `self.ffmpeg()` to
add a counter for the total number of operations remaining in the edit
that currently has priority. Similarly to the `self.saves_in_progress`
counter, this only appears when there are multiple operations to
report (meaning this is irrelevant for concatenation).

A full string would look something like:
`2 edits in progress - Trimming [1/3] (25%)`