Fixed several obscure drag-and-drop cursor bugs
This commit may seem simple, but this took two days of experimentation.
`QVideoPlayer`, `self.dockControls`, and `self.menubar` all now set and
unset their base cursors in a variety of places (context menus, drop
events, enter/leave events) as a hack to fix MANY annoying instances of
the cursor getting stuck as a resize/busy cursor after a drag-and-drop.

A few scenarios where variations of this bug could occur (after a drop):
After mousing off the player, after opening the context menu with
or without right-click, when the window isn't focused, when the
window is fullscreen, when dropping near the edge of the window
too quickly or slowly, etc.

Did I fix all of them? I hope so. Is it perfect? Nope, there are a few
scenarios where the cursor may flicker for a frame or two before
correcting itself, but I'm still happy with it. This solution works
shockingly well, and my earlier attempts over the last two days
involved much worse compromises.