The output text box now expands across the window
Before, Qt-strangeness resulted in `self.lineOutput` not expanding
correctly while `self.frameCropInfo` wasn't visible. Various layouts,
spacings, size policies, and stretches on `self.dockControls` have
been adjusted/simplified to allow the text box to expand fully
(including when the crop frame is visible).

While this ensures filenames are never cut off as long as the window is
wide enough, this also makes dragging the controls to move the window
somewhat awkward since so much space is now devoted to the mostly empty,
invisible (depending on theme) text box. This may be addressed in the
future.