Moved `self.ffmpeg()` to external class
The `SaveParameters` class has been renamed to `Edit`, and many values
that were previously being tracked in `self.ffmpeg()` have become
properties. `self.ffmpeg()` is now called through `Edit.ffmpeg()`,
with many of the same parameters, with the inlined `get_progress_text()`
and `set_progress_bar()` functions becoming separate `Edit` methods that
each edit can call on themselves.

I took quite a while to decide if expanding the new class this way was
actually the direction I wanted to take or not, but I think it'll be
more useful in the future to be able to access these methods and alter
these properties in real time/in a logical way, even if actually setting
up and starting an FFmpeg operation is less intuitive now.