Added context menu/cursor to edit progress bar
`self.editProgressBarContextMenuEvent()` shows the title/progress and
a submenu for cancelling or setting the priority on each active edit
(more actions in the future). Includes an extra "Cancel all" action.

The progress bar now uses the pointing hand cursor (the one used for
links) to make it clear that it can be interacted with while visible.