Improvements to context menu submenus/info actions
All media info actions are now shown in fullscreen mode, since the
contents of your titlebar are irrelevant there. A new setting for
controlling when the menubar submenus appear in the context menu
has been added, `checkContextShowSubmenus`. It's tri-state - you
set the submenus to always appear, never appear, or only appear
when the menubar itself is not visible (the new default).

Additionally, the top section of the settings dialog has been
rearranged and a few titles/tooltips have been changed.