Popups for concating files w/ different dimensions
Files that are about to be concatenated are now probed beforehand (if
possible) to ensure that they're all the same dimensions. If not, a
popup will appear showing the dimensions of each file in the list:
for re-encoding, the popup will explain that they cannot proceed,
but for stream-copying, the popup will warn the user but give them
the option to continue anyway.

If FFprobe is disabled, a third popup appears if the FFmpeg error for
this type of invalid concatenation is detected in its output.

Because the files are now probed before starting, the frame rate/count
hints are now set in `self.concatenate()` and are passed as parameters
to `self._concatenate()`.

Also, `self.concatenate()` now accepts None for the `action` parameter
(defaulting to `self.actionCatDialog`) and properly populates the dialog
when `files` are provided with `self.actionCatDialog` as the `action`.

Additionally, `qthelpers.getPopup()` now includes a `flags`
parameter (used to set `Qt.WindowStaysOnTopHint` for one of
the concatenation warnings).