Edits now auto-update the priority on completion
Instances of `self.reset_save_progress_bar()` have been replaced with a
new method, `self.reset_edit_priority()`. This method is very convoluted
on account of my paranoia about race conditions after dealing with the
high-precision slider for so long. `self.lock_edit_priority` and new
parameters `ignore_lock` and `conditional` have been added to
`Edit.give_priority()` in wake of this madness.

The idea is that we jump the edit closest to completion percentage-wise,
preferring an unpaused one if any exist. If 20+ edits are running (for
some reason), we jump to whatever the first unpaused edit is, otherwise
the first edit in general if everything is paused.

Additionally, `self.get_edit_with_priority()` and
`self.reset_save_progress_bar()` have been moved, with the
latter being renamed to `self.hide_edit_progress()`, for clarity.