Open handle to output path while editing
The output paths for edits are now locked for real using `open()` in
append mode ("a") while editing. The handle is closed immediately after
the first stage of editing using a new helper function `close_handle()`
which also deletes the associated file if nothing was occupying that
path before.

This change has multiple positive effects:
1. `util.get_unique_path()` no longer defaults to filenames that are
   about to be written to
2. It is now more obvious to the user where they can & cannot save
3. It is now more difficult to disrupt PyPlayer while editing