Show popup (but continue) if dest can't be renamed
`self.cleanup_edit()` now shows a popup if it doesn't have permission
to rename the temporary output path to its actual output path. The edit
continues and cleans up normally, but while sticking with the temporary
path. `self.cleanup_edit()` as a result now returns a string containing
the actual "final" destination of the edit if the output file itself
is still valid, or an empty string otherwise. This allows post-edit
activities to continue, such as opening the temporary path like
it's the correct path.