FFmpeg's stdout is now fully parsed in batches
This massively optimizes parsing/updating progress for edits, especially
multiple edits at once. This uses some trickery I couldn't find anywhere
on StackOverflow to seek forwards and backwards in the pipe in order to
only parse output that's ready to be parsed, without having to wait for
buffering. As a result, I'm not sure this will actually work on Linux,
but we'll blow that bridge up when we get to it.

We no longer update and sleep after every single "frame=" line (no idea
why I was doing that), instead tracking the latest frame and only
applying it once we're caught up. Besides optimization, this also fixes
cases where our progress lags behind FFmpeg's actual progress, as well
as cases where we get a janky series of rapid updates after switching
priorities.

Also, now that sleeping is safe, we sleep 0.5 seconds for priority edits
and 1.0 seconds otherwise (split into two sleeps so updates can resume
quickly if needed). Other minor optimizations to sleeping have been done
as well.

Additionally, because we no longer rely on blocking methods, paused
edits now sleep manually for 1 second at a time, refreshing the title
between sleeps if appropriate. This fixes the titlebar not updating
while the priority edit is paused.