Added `QVideoList.generate_thumbnails()`
Replaces `QVideoList.get_thumbnail()` with a more efficient version that
only needs to be called once. Concurrently generates thumbnails in one
thread, 16 at a time. This avoids bottlenecking and crashes caused by
needing to generate a huge number of thumbnails at once.

---

Last commit, I moved `delete_temp_path()` to `qthelpers.py`, expecting
to use it to delete the temporary thumbnails in this commit. Moments
after pushing the commit, I realized it would be better to also delete
the thumbnails concurrently since it's safer and we'll already be in a
thread. I now regret pushing the last commit since it makes far more
sense for `delete_temp_path()` to stay as an isolated utility function
in `main.pyw`.

Oops. Maybe I'll move it back eventually during a random cleanup commit.