Added settings for beeping and pausing after edits
`checkFocusIgnoreFullscreenEditsOnly`, `checkEditOpenRejectedBeep`,
`checkEditFocusRejectedBeep`, and `checkEditFocusRejectedPause` have
been added for playing a notification sound when edits finish without
opening themselves, or playing a sound and/or having the player start
paused when opening a newly finished edit without focusing the window,
and for having the "ignore focus settings if foreground is fullscreen"
setting only apply to edits.

`self.open()` now takes `_from_edit`, `pause_if_focus_rejected`, and
`beep_if_focus_rejected` parameters. The latter two parameters are only
used for edits for now, but I decided on making them as part of a more
generalized API in case I expand on these settings in the future.

The intent for these settings is to not have edits distract you by
stealing focus away (especially from a game) or open themselves in
the background, forcing you to alt-tab to pause them.