Improved color picker, added `util.get_font_path()`
`self.show_color_picker()` now takes `initial_color`, `alpha`, and
`button` parameters to make it more generalized. It also now returns
the selected color if possible, otherwise the starting color (or None).
The starting color is now derived from `button`'s stylesheet if one
is present and `initial_color` is not provided. `alpha` adds an alpha
control to the picker.

`util.get_font_path()` takes a font family and uses the registry to find
the associated filepath, since Qt refuses to expose this information.
It's Windows-only (no idea how to do this on Linux) and is something
that will be used in the next commit before immediately having any
references to it removed shortly after. It will likely never be used
again. I don't know if I'll remove the function entirely though.
It's pretty cool. Maybe we'll just keep it for a while.