Fixed regression with ' vs " in filenames
Back when I ported `sanitize_filename.sanitize()` to `util.py`, I
replaced `"` with `'` around the strings in the blacklist for style
reasons, not noticing that `"` itself was one of the blacklisted
characters. So for quite some time now while renaming, `'` has been
wrongly sanitized (very annoying) and `"` throws an error (which I
somehow never noticed). Oops.