QoL for right-click, delete prompt, & volume boost
Nope. Still not the months-old changes.

- Shift-right-clicking the player now always shows the full context
menu, regardless of your settings.

- The deletion prompt's header has been slightly reworded. It also now
includes both the number of files that are marked for deletion, and the
number of files that are actually checked in the prompt (which updates
in real-time).

- Changing the volume-boost too soon after opening a file (0.35 seconds)
will be assumed to be accidental and will be ignored. To accomplish
this, `self.last_open_time` has been added.