Added data-retaining system for failed edits
Half/half: I did this months ago, but expanded on it today/yesterday.

Adds `self.save_remnants`, a dict for tracking special data from edits
which can be retrieved in case the edit fails (or hasn't finished yet).
This is useful for restoring dialogs and other data so the user can
quickly retry the same or a similar edit.

It's a dictionary with the start time as the key and a nested dictionary
as the value, which always contains at least a `"_in_progress"` key. The
main dictionary is culled every time an edit is successfully completed,
removing all save remnants that started BEFORE that edit and have
`"_in_progress": False`.

Also adds `self.get_save_remnant()`, a convenience method for retrieving
the most recent data for a given key according to several parameters.

Additionally, `self.save()` and `self.save_as()` now return True if they
successfully open `self._save()`, otherwise they return None. This can
be used to detect whether or not a save attempt was cancelled early.