Tracks and delays are now preserved after restart
libVLC does not remember track selections or delays once the player is
stopped. This adds a number of properties and methods for attempting to
safely and quickly restore the previous selections after stopping the
player or re-opening the currently open file.

This has the side-effect of improving the consistency of auto-disabling
subtitle tracks on open.