Added custom player support & QMediaPlayer :)
I don't know why I suddenly embarked on this journey, but it took
several days and was VERY worth it. It's not totally complete and might
undergo some changes before release, but the general idea is there.

The idea is: You define a player that subclasses `PyPlayerBackend`. You
define a few bits of metadata like the name the features it supports,
and then you have free-reign over a variety of events and methods. This
is very much meant to be extendable by users and in the (near!) future,
you'll be able to create and use your own players without recompiling.

All* of libVLC's weird behavior has been isolated and sandboxed into its
own class, with actually surprisingly little performance impact. It
works exactly as before (*an extra bit will be ironed out in one of
the next commits).

That brings us to `QMediaPlayer`. At long last, after 2 years, I've
finally added support for it (as a test, basically). Unfortunately,
since we're still on PyQt5, `QMediaPlayer` kinda sucks. It lacks any
form of track manipulation, and accessing frames for snapshots or
drawing over them without killing performance is far harder than I
expected (snapshotting is done with FFmpeg and I flat out didn't add
marquee support). These issues are fixed with PyQt6, but porting
is also a lot harder than I expected. I will likely leave things as
is for now and create a proper `QMediaPlayer` backend in a few months.

For now, you can toggle between the players in real-time with the new
(and likely not long for this world) "Players" submenu, and enjoy the
bliss of true frame-accuracy.

For now, the `QVideoPlayer` widget is still called `vlc`, and I might
just leave it that way, as a legacy thing (like with `self.video`).
At least for a little while.

---

Some other notable stuff:
- `self.restore()` has been created. This method also fixes issues some
  of its various original implementations had
- `self._update_progress_slot()` has ditched `self.fractional_frame`.
  Also, it now uses `self.frame_override` if set. This was done to fix
  a cosmetic issue while switching from VLC to Qt, but hopefully has no
  negative side effects (if anything, it should have positive ones)
- `set_and_update_progress()` now takes a context value, defined as an
  enum `constants.SetProgressContext`. This is how I reimplemented
  offset values for VLC without `set_and_adjust_and_update_progress()`
  (which is gone forever now, hooray)
- The `radioTextPosition` buttons in the settings dialog now use proper
  numbers instead of VLC's wacky nonsense numbers. A similar story is
  found with track manipulation
- `self.set_playback_speed()` now takes an `increment` parameter
- More stuff I forgot