Added glob-search to output textbox using tab
Pressing tab while `self.lineOutput` is focused will now search for your
entry using `glob.glob()`, adding multiple "*" depending on what was
initially entered. The results are shown as a dropdown in a hidden
`QComboBox` overlayed over the textbox.

Recent searches are saved to a dictionary `self.recent_globs`, and lots
of shenanigans happen behind the scenes to preserve/scroll the dropdown.

Apparently, QComboBox is held together with duct tape and glue, so
this required an unreasonable amount of experimentation to get right.
Despite this, I'm extremely happy with how this turned out.

Additional changes:
- Added `QWidgetPassthrough.focusNextPrevChild()` to properly support
ignoring the tab key. Not the best solution, but I didn't want to use
an event filter.
- `QtGui.QCursor().pos()` corrected to `QtGui.QCursor.pos()`
- Fixed regression in `self.restore_tracks()`